** Move flag deleting to UserRESTService
